+++ app/Http/Controllers/Vehicules/VehiculeUpdateController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Vehicules;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Vehicule\UpdateVehiculeRequest;
+use App\Http\Resources\VehiculeResource;
 use App\Http\Traits\ApiResponse;
 use App\Models\Vehicule;
 use Illuminate\Support\Facades\Storage;
@@ -23,7 +24,9 @@ class VehiculeUpdateController extends Controller
         $data = $request->validated();
 
         if ($request->hasFile('photo')) {
-            Storage::disk('public')->delete($vehicule->photo_path);
+            if (!empty($vehicule->photo_path)) {
+                Storage::disk('public')->delete($vehicule->photo_path);
+            }
             $data['photo_path'] = $request->file('photo')->store('vehicules', 'public');
         }
 
@@ -32,6 +35,6 @@ class VehiculeUpdateController extends Controller
         $vehicule->update($data);
         $vehicule->load(['proprietaire', 'livreurPrincipal']);
 
-        return $this->successResponse($vehicule->fresh(), 'Véhicule mis à jour');
+        return $this->successResponse(VehiculeResource::make($vehicule->fresh()), 'Véhicule mis à jour');
     }
 }
